requirements.txt
# Основные зависимости
python-telegram-bot>=20.0
git+https://github.com/huggingface/transformers@v4.49.0-Gemma-3
huggingface-hub>=0.23.0
faiss-cpu>=1.7.4
torch>=2.2.0
numpy>=1.24.0
tqdm>=4.65.0

# Утилиты
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
nest-asyncio>=1.5.8

# Необходимые библиотеки
requests>=2.31.0
pydantic>=2.4.0 